nuwa/requirements.txt
anybadge==1.14.0
boto3==1.28.49
colorama==0.4.6
faiss_cpu==1.7.4
nest_asyncio==1.5.6
numpy==1.25.0
openai==0.27.8
pickleDB==0.9.2
pytest==7.4.0
Requests==2.31.0
sentence_transformers==2.2.2
google.generativeai == 0.4.0
